pyproject.toml
[tool.poetry]
name = "redis_vectorsearch"
version = "0.1.0"
description = "A demonstration of using RedisVL to do vectorsearch"
authors = ["savynorem <savannah.norem@redis.com>"]
license = "MIT"
repository = "https://github.com/sav-norem/redis_vectorsearch"
packages = [{include = "redisvl_demo", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
redisvl = {git = "https://github.com/sav-norem/redis-vl-python.git"}
pillow = "^10.4.0"
urllib3 = "^2.2.3"
redis = "^5.0.8"
gradio = "^4.44.0"
pandas = "^2.2.2"
sentence-transformers = "^3.1.1"
docker = "^7.1.0"
pytest = "^8.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
